# Conflicts:
#	HSF302_Group3_Project/src/main/java/clothes/hsf302_group3_project/controller/HomePageController.java
#	HSF302_Group3_Project/src/main/java/clothes/hsf302_group3_project/controller/ProductController.java
#	HSF302_Group3_Project/src/main/java/clothes/hsf302_group3_project/controller/ProductSizeController.java
#	HSF302_Group3_Project/src/main/java/clothes/hsf302_group3_project/converter/ConverterDTO.java
#	HSF302_Group3_Project/src/main/java/clothes/hsf302_group3_project/dto/request/GetOrderRequest.java
#	HSF302_Group3_Project/src/main/java/clothes/hsf302_group3_project/dto/request/GetUserRequest.java
#	HSF302_Group3_Project/src/main/java/clothes/hsf302_group3_project/dto/response/CartDTO.java
#	HSF302_Group3_Project/src/main/java/clothes/hsf302_group3_project/dto/response/CartItemDTO.java
#	HSF302_Group3_Project/src/main/java/clothes/hsf302_group3_project/dto/response/ProductSizeDTO.java
#	HSF302_Group3_Project/src/main/java/clothes/hsf302_group3_project/entity/Product.java
#	HSF302_Group3_Project/src/main/java/clothes/hsf302_group3_project/enums/Size.java
#	HSF302_Group3_Project/src/main/resources/application.properties
#	HSF302_Group3_Project/src/main/resources/templates/HomePageProduct/HomePage.html
#	HSF302_Group3_Project/target/classes/application.properties
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/controller/CategoryController.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/controller/HomePageController.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/controller/ProductController.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/converter/ConverterDTO.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/dto/response/CartDTO.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/dto/response/CartItemDTO.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/dto/response/CategoryDTO.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/dto/response/ProductDTO.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/dto/response/ProductSizeDTO.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/dto/response/UserDTO$UserDTOBuilder.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/dto/response/UserDTO.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/entity/CartItem.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/entity/Category.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/entity/Order.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/entity/Product.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/entity/ProductSize.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/enums/OrderStatus.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/exception/GlobalExceptionHandler.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/repository/OrderItemRepository.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/repository/OrderRepository.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/repository/ProductRepository.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/repository/ProductSizeRepository.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/repository/UserRepository.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/security/config/SecurityConfig.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/service/CartService.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/service/Impl/CategoryServiceImpl.class
#	HSF302_Group3_Project/target/classes/clothes/hsf302_group3_project/service/Impl/ProductServiceImpl.class
#	HSF302_Group3_Project/target/classes/templates/admin/customer.html
#	HSF302_Group3_Project/target/classes/templates/admin/fragment/sidebar.html
#	HSF302_Group3_Project/target/classes/templates/admin/order.html
#	HSF302_Group3_Project/target/classes/templates/admin/shipper.html
#	HSF302_Group3_Project/target/classes/templates/cart.html
#	HSF302_Group3_Project/target/classes/templates/checkout.html
#	HSF302_Group3_Project/target/classes/templates/product/create.html
#	HSF302_Group3_Project/target/classes/templates/product/list.html
#	HSF302_Group3_Project/target/classes/templates/product/update.html

+++ HSF302_Group3_Project/src/main/java/clothes/hsf302_group3_project/service/Impl/CategoryServiceImpl.java
@@ -12,7 +12,6 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
-
 @Service
 public class CategoryServiceImpl implements CategoryService {
 
@@ -47,7 +46,7 @@ public class CategoryServiceImpl implements CategoryService {
         Category category = new Category();
         category.setName(categoryDTO.getName());
         Category savedCategory = categoryRepository.save(category);
-        return new CategoryDTO(savedCategory.getId(), savedCategory.getName(), null, 0);
+        return new CategoryDTO(savedCategory.getId(), savedCategory.getName());
     }
 
     @Override
@@ -57,7 +56,7 @@ public class CategoryServiceImpl implements CategoryService {
                 .orElseThrow(() -> new ResourceNotFoundException("Category not found"));
         category.setName(categoryDTO.getName());
         Category updatedCategory = categoryRepository.save(category);
-        return new CategoryDTO(updatedCategory.getId(), updatedCategory.getName(), null, 0);
+        return new CategoryDTO(updatedCategory.getId(), updatedCategory.getName());
     }
 
     @Override
@@ -65,14 +64,17 @@ public class CategoryServiceImpl implements CategoryService {
     public void deleteCategory(Integer id) {
         Category category = categoryRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Category not found"));
-        categoryRepository.delete(category);
+        if(category.getProducts().size() <=0) {
+            categoryRepository.delete(category);
+        }
+        else throw new ResourceNotFoundException("Product in category already exists");
     }
 
     @Override
     public CategoryDTO getCategoryById(Integer id) {
         Category category = categoryRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Category not found"));
-        return new CategoryDTO(category.getId(), category.getName(), null, 0);
+        return new CategoryDTO(category.getId(), category.getName());
     }
